+++ contracts/hackatom/src/lib.rs
@@ -16,7 +16,7 @@ mod wasm {
     #[no_mangle]
     pub extern "C" fn init(params_ptr: *mut c_void, msg_ptr: *mut c_void) -> *mut c_void {
         exports::do_init(
-            &contract::init::<imports::ExternalStorage>,
+            &contract::init::<imports::ExternalStorage, imports::ExternalApi>,
             params_ptr,
             msg_ptr,
         )
@@ -25,7 +25,7 @@ mod wasm {
     #[no_mangle]
     pub extern "C" fn handle(params_ptr: *mut c_void, msg_ptr: *mut c_void) -> *mut c_void {
         exports::do_handle(
-            &contract::handle::<imports::ExternalStorage>,
+            &contract::handle::<imports::ExternalStorage, imports::ExternalApi>,
             params_ptr,
             msg_ptr,
         )
@@ -33,6 +33,9 @@ mod wasm {
 
     #[no_mangle]
     pub extern "C" fn query(msg_ptr: *mut c_void) -> *mut c_void {
-        exports::do_query(&contract::query::<imports::ExternalStorage>, msg_ptr)
+        exports::do_query(
+            &contract::query::<imports::ExternalStorage, imports::ExternalApi>,
+            msg_ptr,
+        )
     }
 }