+++ lib/vm/src/cache.rs
@@ -4,7 +4,7 @@ use std::path::PathBuf;
 use lru::LruCache;
 use snafu::ResultExt;
 
-use cosmwasm::storage::Storage;
+use cosmwasm::traits::{Api, Extern, Storage};
 
 use crate::backends::{backend, compile};
 use crate::errors::{Error, IntegrityErr, IoErr};
@@ -15,17 +15,24 @@ use crate::wasm_store::{load, save, wasm_hash};
 static WASM_DIR: &str = "wasm";
 static MODULES_DIR: &str = "modules";
 
-pub struct CosmCache<T: Storage + 'static> {
+pub struct CosmCache<S: Storage + 'static, A: Api + 'static> {
     wasm_path: PathBuf,
     modules: FileSystemCache,
-    instances: Option<LruCache<WasmHash, Instance<T>>>,
+    instances: Option<LruCache<WasmHash, Instance<S, A>>>,
 }
 
-impl<T> CosmCache<T>
+impl<S, A> CosmCache<S, A>
 where
-    T: Storage + 'static,
+    S: Storage + 'static,
+    A: Api + 'static,
 {
     /// new stores the data for cache under base_dir
+    ///
+    /// # Safety
+    ///
+    /// This function is marked unsafe due to `FileSystemCache::new`, which implicitly
+    /// assumes the disk contents are correct, and there's no way to ensure the artifacts
+    //  stored in the cache haven't been corrupted or tampered with.
     pub unsafe fn new<P: Into<PathBuf>>(base_dir: P, cache_size: usize) -> Result<Self, Error> {
         let base = base_dir.into();
         let wasm_path = base.join(WASM_DIR);
@@ -64,14 +71,14 @@ where
     }
 
     /// get instance returns a wasmer Instance tied to a previously saved wasm
-    pub fn get_instance(&mut self, id: &[u8], storage: T) -> Result<Instance<T>, Error> {
+    pub fn get_instance(&mut self, id: &[u8], deps: Extern<S, A>) -> Result<Instance<S, A>, Error> {
         let hash = WasmHash::generate(&id);
 
         // pop from lru cache if present
         if let Some(cache) = &mut self.instances {
             let val = cache.pop(&hash);
             if let Some(inst) = val {
-                inst.leave_storage(Some(storage));
+                inst.leave_storage(Some(deps.storage));
                 return Ok(inst);
             }
         }
@@ -79,23 +86,25 @@ where
         // try from the module cache
         let res = self.modules.load_with_backend(hash, backend());
         if let Ok(module) = res {
-            return Instance::from_module(&module, storage);
+            return Instance::from_module(&module, deps);
         }
 
         // fall back to wasm cache (and re-compiling) - this is for backends that don't support serialization
         let wasm = self.load_wasm(id)?;
-        Instance::from_code(&wasm, storage)
+        Instance::from_code(&wasm, deps)
     }
 
-    pub fn store_instance(&mut self, id: &[u8], instance: Instance<T>) -> Option<T> {
+    pub fn store_instance(&mut self, id: &[u8], instance: Instance<S, A>) -> Option<Extern<S, A>> {
         if let Some(cache) = &mut self.instances {
             let hash = WasmHash::generate(&id);
             let storage = instance.take_storage();
+            let api = instance.api; // copy it
             cache.put(hash, instance);
-            storage
-        } else {
-            None
+            if let Some(storage) = storage {
+                return Some(Extern { storage, api });
+            }
         }
+        None
     }
 }
 
@@ -105,8 +114,8 @@ mod test {
     use tempfile::TempDir;
 
     use crate::calls::{call_handle, call_init};
-    use cosmwasm::mock::MockStorage;
-    use cosmwasm::types::{coin, mock_params};
+    use cosmwasm::mock::{dependencies, mock_params};
+    use cosmwasm::types::coin;
 
     static CONTRACT: &[u8] = include_bytes!("../testdata/contract.wasm");
 
@@ -115,11 +124,11 @@ mod test {
         let tmp_dir = TempDir::new().unwrap();
         let mut cache = unsafe { CosmCache::new(tmp_dir.path(), 10).unwrap() };
         let id = cache.save_wasm(CONTRACT).unwrap();
-        let storage = MockStorage::new();
-        let mut instance = cache.get_instance(&id, storage).unwrap();
+        let deps = dependencies(20);
+        let mut instance = cache.get_instance(&id, deps).unwrap();
 
         // run contract
-        let params = mock_params("creator", &coin("1000", "earth"), &[]);
+        let params = mock_params(&instance.api, "creator", &coin("1000", "earth"), &[]);
         let msg = r#"{"verifier": "verifies", "beneficiary": "benefits"}"#.as_bytes();
 
         // call and check
@@ -133,18 +142,23 @@ mod test {
         let tmp_dir = TempDir::new().unwrap();
         let mut cache = unsafe { CosmCache::new(tmp_dir.path(), 10).unwrap() };
         let id = cache.save_wasm(CONTRACT).unwrap();
-        let storage = MockStorage::new();
-        let mut instance = cache.get_instance(&id, storage).unwrap();
+        let deps = dependencies(20);
+        let mut instance = cache.get_instance(&id, deps).unwrap();
 
         // init contract
-        let params = mock_params("creator", &coin("1000", "earth"), &[]);
+        let params = mock_params(&instance.api, "creator", &coin("1000", "earth"), &[]);
         let msg = r#"{"verifier": "verifies", "beneficiary": "benefits"}"#.as_bytes();
         let res = call_init(&mut instance, &params, msg).unwrap();
         let msgs = res.unwrap().messages;
         assert_eq!(msgs.len(), 0);
 
         // run contract - just sanity check - results validate in contract unit tests
-        let params = mock_params("verifies", &coin("15", "earth"), &coin("1015", "earth"));
+        let params = mock_params(
+            &instance.api,
+            "verifies",
+            &coin("15", "earth"),
+            &coin("1015", "earth"),
+        );
         let msg = b"{}";
         let res = call_handle(&mut instance, &params, msg).unwrap();
         let msgs = res.unwrap().messages;
@@ -158,30 +172,35 @@ mod test {
         let id = cache.save_wasm(CONTRACT).unwrap();
 
         // these differentiate the two instances of the same contract
-        let storage1 = MockStorage::new();
-        let storage2 = MockStorage::new();
+        let deps1 = dependencies(20);
+        let deps2 = dependencies(20);
 
         // init instance 1
-        let mut instance = cache.get_instance(&id, storage1).unwrap();
-        let params = mock_params("owner1", &coin("1000", "earth"), &[]);
+        let mut instance = cache.get_instance(&id, deps1).unwrap();
+        let params = mock_params(&instance.api, "owner1", &coin("1000", "earth"), &[]);
         let msg = r#"{"verifier": "sue", "beneficiary": "mary"}"#.as_bytes();
         let res = call_init(&mut instance, &params, msg).unwrap();
         let msgs = res.unwrap().messages;
         assert_eq!(msgs.len(), 0);
-        let storage1 = cache.store_instance(&id, instance).unwrap();
+        let deps1 = cache.store_instance(&id, instance).unwrap();
 
         // init instance 2
-        let mut instance = cache.get_instance(&id, storage2).unwrap();
-        let params = mock_params("owner2", &coin("500", "earth"), &[]);
+        let mut instance = cache.get_instance(&id, deps2).unwrap();
+        let params = mock_params(&instance.api, "owner2", &coin("500", "earth"), &[]);
         let msg = r#"{"verifier": "bob", "beneficiary": "john"}"#.as_bytes();
         let res = call_init(&mut instance, &params, msg).unwrap();
         let msgs = res.unwrap().messages;
         assert_eq!(msgs.len(), 0);
-        let storage2 = cache.store_instance(&id, instance).unwrap();
+        let deps2 = cache.store_instance(&id, instance).unwrap();
 
         // run contract 2 - just sanity check - results validate in contract unit tests
-        let mut instance = cache.get_instance(&id, storage2).unwrap();
-        let params = mock_params("bob", &coin("15", "earth"), &coin("1015", "earth"));
+        let mut instance = cache.get_instance(&id, deps2).unwrap();
+        let params = mock_params(
+            &instance.api,
+            "bob",
+            &coin("15", "earth"),
+            &coin("1015", "earth"),
+        );
         let msg = b"{}";
         let res = call_handle(&mut instance, &params, msg).unwrap();
         let msgs = res.unwrap().messages;
@@ -189,8 +208,13 @@ mod test {
         let _ = cache.store_instance(&id, instance).unwrap();
 
         // run contract 1 - just sanity check - results validate in contract unit tests
-        let mut instance = cache.get_instance(&id, storage1).unwrap();
-        let params = mock_params("sue", &coin("15", "earth"), &coin("1015", "earth"));
+        let mut instance = cache.get_instance(&id, deps1).unwrap();
+        let params = mock_params(
+            &instance.api,
+            "sue",
+            &coin("15", "earth"),
+            &coin("1015", "earth"),
+        );
         let msg = b"{}";
         let res = call_handle(&mut instance, &params, msg).unwrap();
         let msgs = res.unwrap().messages;