+++ lib/vm/src/calls.rs
@@ -1,14 +1,14 @@
 use snafu::ResultExt;
 
 use cosmwasm::serde::{from_slice, to_vec};
-use cosmwasm::storage::Storage;
+use cosmwasm::traits::{Api, Storage};
 use cosmwasm::types::{ContractResult, Params, QueryResult};
 
 use crate::errors::{Error, ParseErr, RuntimeErr, SerializeErr};
 use crate::instance::{Func, Instance};
 
-pub fn call_init<T: Storage + 'static>(
-    instance: &mut Instance<T>,
+pub fn call_init<S: Storage + 'static, A: Api + 'static>(
+    instance: &mut Instance<S, A>,
     params: &Params,
     msg: &[u8],
 ) -> Result<ContractResult, Error> {
@@ -18,8 +18,8 @@ pub fn call_init<T: Storage + 'static>(
     Ok(res)
 }
 
-pub fn call_handle<T: Storage + 'static>(
-    instance: &mut Instance<T>,
+pub fn call_handle<S: Storage + 'static, A: Api + 'static>(
+    instance: &mut Instance<S, A>,
     params: &Params,
     msg: &[u8],
 ) -> Result<ContractResult, Error> {
@@ -29,8 +29,8 @@ pub fn call_handle<T: Storage + 'static>(
     Ok(res)
 }
 
-pub fn call_query<T: Storage + 'static>(
-    instance: &mut Instance<T>,
+pub fn call_query<S: Storage + 'static, A: Api + 'static>(
+    instance: &mut Instance<S, A>,
     msg: &[u8],
 ) -> Result<QueryResult, Error> {
     let data = call_query_raw(instance, msg)?;
@@ -38,13 +38,13 @@ pub fn call_query<T: Storage + 'static>(
     Ok(res)
 }
 
-pub fn call_query_raw<T: Storage + 'static>(
-    instance: &mut Instance<T>,
+pub fn call_query_raw<S: Storage + 'static, A: Api + 'static>(
+    instance: &mut Instance<S, A>,
     msg: &[u8],
 ) -> Result<Vec<u8>, Error> {
     // we cannot resuse the call_raw functionality as it assumes a param variable... just do it inline
     let msg_offset = instance.allocate(msg)?;
-    let func: Func<(u32), (u32)> = instance.func("query")?;
+    let func: Func<u32, u32> = instance.func("query")?;
     let res_offset = func.call(msg_offset).context(RuntimeErr {})?;
     let data = instance.memory(res_offset);
     // free return value in wasm (arguments were freed in wasm code)
@@ -52,24 +52,24 @@ pub fn call_query_raw<T: Storage + 'static>(
     Ok(data)
 }
 
-pub fn call_init_raw<T: Storage + 'static>(
-    instance: &mut Instance<T>,
+pub fn call_init_raw<S: Storage + 'static, A: Api + 'static>(
+    instance: &mut Instance<S, A>,
     params: &[u8],
     msg: &[u8],
 ) -> Result<Vec<u8>, Error> {
     call_raw(instance, "init", params, msg)
 }
 
-pub fn call_handle_raw<T: Storage + 'static>(
-    instance: &mut Instance<T>,
+pub fn call_handle_raw<S: Storage + 'static, A: Api + 'static>(
+    instance: &mut Instance<S, A>,
     params: &[u8],
     msg: &[u8],
 ) -> Result<Vec<u8>, Error> {
     call_raw(instance, "handle", params, msg)
 }
 
-fn call_raw<T: Storage + 'static>(
-    instance: &mut Instance<T>,
+fn call_raw<S: Storage + 'static, A: Api + 'static>(
+    instance: &mut Instance<S, A>,
     name: &str,
     params: &[u8],
     msg: &[u8],
@@ -77,7 +77,7 @@ fn call_raw<T: Storage + 'static>(
     let param_offset = instance.allocate(params)?;
     let msg_offset = instance.allocate(msg)?;
 
-    let func: Func<(u32, u32), (u32)> = instance.func(name)?;
+    let func: Func<(u32, u32), u32> = instance.func(name)?;
     let res_offset = func.call(param_offset, msg_offset).context(RuntimeErr {})?;
 
     let data = instance.memory(res_offset);