+++ src/imports.rs
@@ -1,17 +1,41 @@
 #![cfg(target_arch = "wasm32")]
 
 use std::ffi::c_void;
+use std::str::from_utf8;
 use std::vec::Vec;
 
+use snafu::ResultExt;
+
+use crate::errors::{ContractErr, Result, Utf8Err};
 use crate::memory::{alloc, build_slice, consume_slice, Slice};
-use crate::storage::Storage;
+use crate::traits::{Api, Extern, Storage};
 
 // this is the buffer we pre-allocate in get - we should configure this somehow later
 static MAX_READ: usize = 2000;
 
+// this should be plenty for any address representation
+static ADDR_BUFFER: usize = 72;
+
+// TODO: use feature switches to enable precompile dependencies in the future,
+// so contracts that need less
 extern "C" {
+    // these are needed for storage
     fn c_read(key: *const c_void, value: *mut c_void) -> i32;
     fn c_write(key: *const c_void, value: *mut c_void);
+
+    // we define two more functions that must be available...
+    // they take a string and return to a preallocated buffer
+    // returns negative on error, length of returned data on success
+    fn c_canonical_address(human: *const c_void, canonical: *mut c_void) -> i32;
+    fn c_human_address(canonical: *const c_void, human: *mut c_void) -> i32;
+}
+
+// dependencies are all external requirements that can be injected in a real-wasm contract
+pub fn dependencies() -> Extern<ExternalStorage, ExternalApi> {
+    Extern {
+        storage: ExternalStorage::new(),
+        api: ExternalApi::new(),
+    }
 }
 
 #[derive(Clone)]
@@ -52,5 +76,53 @@ impl Storage for ExternalStorage {
         unsafe {
             c_write(key_ptr, value_ptr);
         }
+    }
+}
+
+#[derive(Copy, Clone)]
+pub struct ExternalApi {}
+
+impl ExternalApi {
+    pub fn new() -> ExternalApi {
+        ExternalApi {}
+    }
+}
+
+impl Api for ExternalApi {
+    fn canonical_address(&self, human: &str) -> Result<Vec<u8>> {
+        let send = build_slice(human.as_bytes());
+        let send_ptr = &*send as *const Slice as *const c_void;
+        let canon = alloc(ADDR_BUFFER);
+
+        let read = unsafe { c_canonical_address(send_ptr, canon) };
+        if read < 0 {
+            return ContractErr {
+                msg: "canonical_address returned error",
+            }
+            .fail();
+        }
+
+        let mut out = unsafe { consume_slice(canon)? };
+        out.truncate(read as usize);
+        Ok(out)
+    }
+
+    fn human_address(&self, canonical: &[u8]) -> Result<String> {
+        let send = build_slice(canonical);
+        let send_ptr = &*send as *const Slice as *const c_void;
+        let human = alloc(ADDR_BUFFER);
+
+        let read = unsafe { c_human_address(send_ptr, human) };
+        if read < 0 {
+            return ContractErr {
+                msg: "humanize_address returned error",
+            }
+            .fail();
+        }
+
+        let mut out = unsafe { consume_slice(human)? };
+        out.truncate(read as usize);
+        let result = from_utf8(&out).context(Utf8Err {})?.to_string();
+        Ok(result)
     }
 }